Move logger module to logging
* Since we document more than one class the change makes sense
