+++ hoomd/__init__.py
@@ -72,13 +72,12 @@ from hoomd.simulation import Simulation
 from hoomd.state import State
 from hoomd.operations import Operations
 from hoomd.snapshot import Snapshot
-from hoomd.logger import Logger
 from hoomd.box import Box
 from hoomd import tuner
 from hoomd.custom_action import CustomAction, ParticleDataFlags
 from hoomd.custom_operation import _CustomOperation
 
-from hoomd import logger
+from hoomd import logging
 
 from hoomd._hoomd import WalltimeLimitReached
