+++ hoomd/hpmc/update.py
@@ -7,7 +7,7 @@
 from . import _hpmc
 from . import integrate
 from hoomd import _hoomd
-from hoomd.logger import Loggable
+from hoomd.logging import Loggable
 from hoomd.update import _updater
 from hoomd.operation import _Updater
 from hoomd.parameterdicts import ParameterDict
