+++ hoomd/operation.py
@@ -13,7 +13,7 @@ from hoomd.util import is_iterable, dict_map, str_to_tuple_keys
 from hoomd.trigger import Periodic, Trigger
 from hoomd.variant import Variant, Constant
 from hoomd.filter import _ParticleFilter
-from hoomd.logger import Loggable
+from hoomd.logging import Loggable
 from hoomd.typeconverter import RequiredArg
 from hoomd.util import NamespaceDict
 from hoomd._hoomd import GSDStateReader, is_MPI_available
