+++ hoomd/pytest/CMakeLists.txt
@@ -9,7 +9,7 @@ set(files __init__.py
           test_operation.py
           test_syncedlist.py
           test_triggeredops.py
-          test_logger.py
+          test_logging.py
           test_filter.py
           dummy.py
           test_snapshot.py
