+++ hoomd/pytest/test_logging.py
@@ -1,7 +1,7 @@
 from pytest import raises, fixture
-from hoomd.logger import LoggerQuantity, SafeNamespaceDict, Logger, dict_map
-from hoomd.logger import generate_namespace
-from hoomd.logger import Loggable
+from hoomd.logging import (
+    LoggerQuantity, SafeNamespaceDict, Logger,
+    dict_map, generate_namespace, Loggable)
 
 
 class DummyNamespace:
@@ -10,7 +10,7 @@ class DummyNamespace:
 
 @fixture(scope='module')
 def dummy_namespace():
-    return ('pytest', 'test_logger', 'DummyNamespace')
+    return ('pytest', 'test_logging', 'DummyNamespace')
 
 
 # ------- Test LoggerQuantity
@@ -36,10 +36,10 @@ class TestLoggerQuantity:
                 break
 
 
-# ------- Test generate_logger_decorator and associated functions
+# ------- Test generate_logging_decorator and associated functions
 def test_generate_namespace():
     assert generate_namespace(TestLoggerQuantity) == \
-        ('pytest', 'test_logger', 'TestLoggerQuantity')
+        ('pytest', 'test_logging', 'TestLoggerQuantity')
 
 
 class DummyLoggable(metaclass=Loggable):
@@ -187,7 +187,6 @@ def log_quantity():
 @fixture
 def logged_obj():
     return DummyLoggable()
-
 
 @fixture
 def base_namespace():
