+++ hoomd/simulation.py
@@ -146,7 +146,7 @@ class Simulation:
         log data to a file or when using an NPT integrator). Set
         ``always_compute_pressure`` to True to make the per particle virial,
         net virial, and system pressure available to query any time by property
-        or through the :py:class:`hoomd.Logger` interface.
+        or through the :py:class:`hoomd.logging.Logger` interface.
 
         Note:
             Enabling this flag will result in a moderate performance penalty
