+++ sphinx-doc/module-hoomd-logging.rst
@@ -1,16 +1,19 @@
-hoomd.logger
+hoomd.logging
 --------------
 
 .. rubric:: Overview
 
+.. py:currentmodule:: hoomd.logging
+
 .. autosummary::
     :nosignatures:
 
-    hoomd.logger.LoggerQuantity
+    Logger
+    LoggerQuantity
 
 .. rubric:: Details
 
-.. automodule:: hoomd.logger
+.. automodule:: hoomd.logging
     :synopsis: Classes for logging data.
-    :members: LoggerQuantity
+    :members: Logger, LoggerQuantity
     :undoc-members:
